Extract useLogin hook
